Extract bank payment offer lifecycle from contact composition
